Split screenshot export module
